src/image-analysis-service/requirements.txt
torch
requests
confluent-kafka
Pillow
lime
scikit-image
numpy
matplotlib
transformers
accelerate
datasets
torchvision
